Split listing and creation endpoint out of views
